\name{pkgEx-package}
\alias{pkgEx-package}
\alias{pkgEx}
\docType{package}
\title{
\packageTitle{pkgEx}
}
\description{
\packageDescription{pkgEx}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pkgEx}
\packageIndices{pkgEx}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pkgEx}

Maintainer: \packageMaintainer{pkgEx}
}
\keyword{ package }
